Consolidate small utility functions
